feat: comment mentions (#9071)
* feat: autocomplete mentions on comments

* comment model and service, save chart_uuid

* Update packages/backend/src/models/CommentModel/CommentModel.ts


* fix db

---------
