+++ packages/backend/src/models/DashboardModel/DashboardModel.ts
@@ -885,196 +885,4 @@ export class DashboardModel {
             .distinctOn(`${DashboardVersionsTableName}.dashboard_id`)
             .where(`${ProjectTableName}.project_uuid`, projectUuid);
     }
-
-    private async checkDashboardTileExistsInDashboard(
-        dashboardUuid: string,
-        dashboardTileUuid: string,
-    ) {
-        // NOTE: ensure that this dashboard actually contains the tile, since tile uuids might not be unique across dashboards
-        const dashboardTile = await this.database(DashboardTilesTableName)
-            .join(
-                DashboardVersionsTableName,
-                `${DashboardVersionsTableName}.dashboard_version_id`,
-                '=',
-                `${DashboardTilesTableName}.dashboard_version_id`,
-            )
-            .join(
-                DashboardsTableName,
-                `${DashboardsTableName}.dashboard_id`,
-                '=',
-                `${DashboardVersionsTableName}.dashboard_id`,
-            )
-            .where(`${DashboardsTableName}.dashboard_uuid`, dashboardUuid)
-            .andWhere(
-                `${DashboardTilesTableName}.dashboard_tile_uuid`,
-                dashboardTileUuid,
-            )
-            .select(`${DashboardTilesTableName}.dashboard_tile_uuid`)
-            .first();
-
-        if (!dashboardTile) {
-            throw new NotFoundError('Dashboard tile not found for dashboard');
-        }
-    }
-
-    private static parseComments(
-        commentsWithUsers: (DbDashboardTileComments &
-            Pick<DbUser, 'first_name' | 'last_name'>)[],
-        userUuid: string,
-        canUserRemoveAnyComment: boolean,
-    ) {
-        const commentsPerDashboardTile: Record<string, Comment[]> = {}; // Stores comments grouped by their dashboard_tile_uuid
-        const allComments: Record<string, Comment> = {}; // Fast access lookup for parent comments
-        const orphanReplies: Record<string, Comment[]> = {}; // Stores orphan replies keyed by their intended parent's commentId
-
-        commentsWithUsers.forEach((comment) => {
-            const uuid = comment.dashboard_tile_uuid;
-            if (!commentsPerDashboardTile[uuid]) {
-                commentsPerDashboardTile[uuid] = [];
-            }
-
-            const structuredComment: Comment = {
-                commentId: comment.comment_id,
-                text: comment.text,
-                replyTo: comment.reply_to ?? undefined,
-                user: { name: `${comment.first_name} ${comment.last_name}` },
-                createdAt: comment.created_at,
-                resolved: comment.resolved,
-                replies: [],
-                canRemove:
-                    canUserRemoveAnyComment || comment.user_uuid === userUuid,
-            };
-
-            // Directly attach to parent if it's a reply and the parent exists
-            if (
-                structuredComment.replyTo &&
-                allComments[structuredComment.replyTo]
-            ) {
-                allComments[structuredComment.replyTo].replies?.push(
-                    structuredComment,
-                );
-            } else {
-                if (!structuredComment.replyTo) {
-                    // For comments that are not replies, add them to the list
-                    commentsPerDashboardTile[uuid].push(structuredComment);
-                }
-                // Store the comment for future reference
-                allComments[structuredComment.commentId] = structuredComment;
-            }
-
-            // Add the orphan replies to their intended parent if it exists
-            if (orphanReplies[structuredComment.commentId]) {
-                orphanReplies[structuredComment.commentId].forEach(
-                    (orphanReply) => {
-                        structuredComment.replies?.push(orphanReply);
-                    },
-                );
-
-                delete orphanReplies[structuredComment.commentId];
-            }
-
-            // If the comment that this reply is intended for doesn't exist yet, store it as an orphan
-            if (
-                structuredComment.replyTo &&
-                !allComments[structuredComment.replyTo]
-            ) {
-                if (!orphanReplies[structuredComment.replyTo]) {
-                    orphanReplies[structuredComment.replyTo] = [];
-                }
-                orphanReplies[structuredComment.replyTo].push(
-                    structuredComment,
-                );
-            }
-        });
-
-        return commentsPerDashboardTile;
-    }
-
-    async createComment(
-        dashboardUuid: string,
-        dashboardTileUuid: string,
-        text: string,
-        replyTo: string | null,
-        user: LightdashUser,
-    ): Promise<string> {
-        await this.checkDashboardTileExistsInDashboard(
-            dashboardUuid,
-            dashboardTileUuid,
-        );
-        const [{ comment_id: commentId }] = await this.database(
-            DashboardTileCommentsTableName,
-        )
-            .insert({
-                text,
-                dashboard_tile_uuid: dashboardTileUuid,
-                reply_to: replyTo ?? null,
-                user_uuid: user.userUuid,
-            })
-            .returning('comment_id');
-
-        return commentId;
-    }
-
-    async findCommentsForDashboard(
-        dashboardUuid: string,
-        userUuid: string,
-        canUserRemoveAnyComment: boolean,
-    ): Promise<Record<string, Comment[]>> {
-        const dashboard = await this.getById(dashboardUuid);
-
-        const tileUuids = dashboard.tiles.map((tile) => tile.uuid);
-
-        const commentsWithUsers = await this.database(
-            DashboardTileCommentsTableName,
-        )
-            .leftJoin(
-                UserTableName,
-                `${DashboardTileCommentsTableName}.user_uuid`,
-                '=',
-                `${UserTableName}.user_uuid`,
-            )
-            .select<
-                (DbDashboardTileComments &
-                    Pick<DbUser, 'first_name' | 'last_name'>)[]
-            >(
-                `${DashboardTileCommentsTableName}.*`,
-                `${UserTableName}.first_name`,
-                `${UserTableName}.last_name`,
-                `${DashboardTileCommentsTableName}.dashboard_tile_uuid`,
-            )
-            .whereIn(
-                `${DashboardTileCommentsTableName}.dashboard_tile_uuid`,
-                tileUuids,
-            )
-            .andWhere(`${DashboardTileCommentsTableName}.resolved`, false)
-            .orderBy(`${DashboardTileCommentsTableName}.created_at`, 'asc');
-
-        return DashboardModel.parseComments(
-            commentsWithUsers,
-            userUuid,
-            canUserRemoveAnyComment,
-        );
-    }
-
-    async resolveComment(commentId: string): Promise<void> {
-        await this.database(DashboardTileCommentsTableName)
-            .update({ resolved: true })
-            .where('comment_id', commentId);
-    }
-
-    async getCommentOwner(commentId: string): Promise<string | null> {
-        const result = await this.database(DashboardTileCommentsTableName)
-            .select('user_uuid')
-            .where('comment_id', commentId)
-            .first();
-
-        return result ? result.user_uuid : null;
-    }
-
-    async deleteComment(commentId: string): Promise<void> {
-        await this.database(DashboardTileCommentsTableName)
-            .delete()
-            .where('reply_to', commentId)
-            .orWhere('comment_id', commentId);
-    }
 }
